+++ database/repo/table.go
@@ -14,8 +14,8 @@ const (
 CREATE TABLE
 IF NOT EXISTS
 repos (
-	id               SERIAL PRIMARY KEY,
-	user_id          INTEGER,
+	id               BIGSERIAL PRIMARY KEY,
+	user_id          BIGINT,
 	hash             VARCHAR(500),
 	org              VARCHAR(250),
 	name             VARCHAR(250),
@@ -24,19 +24,19 @@ repos (
 	clone            VARCHAR(1000),
 	branch           VARCHAR(250),
 	topics           VARCHAR(1020),
-	build_limit      INTEGER,
-	timeout          INTEGER,
-	counter          INTEGER,
+	build_limit      BIGINT,
+	timeout          BIGINT,
+	counter          BIGINT,
 	visibility       TEXT,
 	private          BOOLEAN,
 	trusted          BOOLEAN,
 	active           BOOLEAN,
-	allow_events     INTEGER,
+	allow_events     BIGINT,
 	pipeline_type    TEXT,
 	previous_name    VARCHAR(100),
 	approve_build    VARCHAR(20),
-	approval_timeout INTEGER,
-	install_id       INTEGER,
+	approval_timeout BIGINT,
+	install_id       BIGINT,
 	UNIQUE(full_name)
 );
 `
